fix: 엔티티 JPA 타입 ↔ DB 컬럼 타입 정합성 + Redis 포트 충돌 회피
## 배경
실제 bootRun 시 Hibernate schema-validation 실패 6건 + Redis 연결 실패 1건 발견.
통합 검증을 통해 수정.

## 엔티티 타입 정합성
DB 컬럼 타입 ↔ Java 필드 불일치 수정. Hibernate validate 통과 확인.

| 엔티티 | 필드 | DB 타입 | Before | After |
| --- | --- | --- | --- | --- |
| User | ext_id | CHAR(26) | length=26 (→VARCHAR) | columnDefinition="char(26)" |
| User | email | VARBINARY(256) | (length 미지정) | columnDefinition="varbinary(256)" |
| User | email_hash | CHAR(64) | length=64 | columnDefinition="char(64)" |
| Gym/Route/ClimbingSession/MediaAsset/FeedPost/Comment | ext_id | CHAR(26) | length=26 | columnDefinition="char(26)" |
| Gym | setting_cycle_days | SMALLINT | Integer | Short |
| ClimbingSession | duration_min | SMALLINT | Integer | Short |
| ClimbingSession | condition | TINYINT | Integer | Byte |
| SessionAttempt | attempts | SMALLINT | Integer | Short |
| Profile | level_self | TINYINT UNSIGNED | Integer | Byte |
| PostMedia | seq | SMALLINT | Integer | Short |
| FeedPost | like_count / comment_count | INT UNSIGNED | Long | Integer |

## 변환기
`CodeEnumConverter<E, Integer>` → `CodeEnumConverter<E, Byte>`
(DB TINYINT ↔ enum code 매핑, 상속 6종 모두 자동 적용)

## Redis 포트 변경 (13306 MySQL 때와 동일 패턴)
로컬 Homebrew Redis(6379)와 충돌 → Docker Redis 를 16379 매핑.

- docker-compose.yml: `6379:6379` → `16379:6379`
- application.yml: REDIS_PORT 기본값 6379 → 16379
- infra/README.md: 접속 정보 + 포트 충돌 회피 주석

## 검증
- `./gradlew :crimp-api:compileJava` SUCCESS
- `docker compose up -d` → `redis-cli -p 16379 ping` PONG
- `./gradlew :crimp-api:bootRun` → Started CrimpApiApplication in 2.86s, Tomcat on 8080
- `GET /actuator/health` → 200 `{"status":"UP"}` (Redis 연결 확인)

## 주의
- 이 PR 은 `develop` 기준이라 PR #15(Auth) 가 포함돼있지 않음.
  auth 엔드포인트 동작 검증은 PR #15 머지 후 가능 (현재는 Spring Security 기본 체인이 401 반환).

+++ api/crimp-core/src/main/java/io/crimp/core/entity/feed/PostMedia.java
@@ -22,11 +22,11 @@ public class PostMedia {
     private PostMediaId id;
 
     @Column(name = "seq", nullable = false)
-    private Integer seq;
+    private Short seq;
 
     private PostMedia(PostMediaId id, int seq) {
         this.id = id;
-        this.seq = seq;
+        this.seq = (short) seq;
     }
 
     public static PostMedia attach(Long postId, Long mediaId, int seq) {
